+++ packages/lightspark-sdk/examples/node-scripts/authHelpers.ts
@@ -5,7 +5,8 @@ export const getCredentialsFromEnvOrThrow = () => {
   const apiTokenClientSecret =
     process.env["LIGHTSPARK_API_TOKEN_CLIENT_SECRET"];
   const nodePassword = process.env["LIGHTSPARK_TEST_NODE_PASSWORD"];
-  const baseUrl = process.env["LIGHTSPARK_EXAMPLE_BASE_URL"] || "api.lightspark.com";
+  const baseUrl =
+    process.env["LIGHTSPARK_EXAMPLE_BASE_URL"] || "api.lightspark.com";
   if (!apiTokenClientId || !apiTokenClientSecret || !nodePassword) {
     throw new Error(
       "Missing test credentials. Please set LIGHTSPARK_API_TOKEN_CLIENT_ID, LIGHTSPARK_API_TOKEN_CLIENT_SECRET, and LIGHTSPARK_TEST_NODE_PASSWORD environment variables."
@@ -15,6 +16,6 @@ export const getCredentialsFromEnvOrThrow = () => {
     apiTokenClientId,
     apiTokenClientSecret,
     nodePassword,
-    baseUrl
+    baseUrl,
   };
 };