+++ packages/wallet-sdk/examples/node-scripts/example.ts
@@ -1,5 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
+import { KeyOrAlias } from "@lightsparkdev/core";
 import {
   CurrencyAmount,
   getTransactionQuery,
@@ -173,14 +174,14 @@ if (!signingPrivateKey) {
     "Missing signing key in the environment. Please set LIGHTSPARK_WALLET_PUB_KEY and LIGHTSPARK_WALLET_PRIV_KEY."
   );
 }
-await client.loadWalletSigningKey(signingPrivateKey);
+await client.loadWalletSigningKey(KeyOrAlias.key(signingPrivateKey));
 console.log(`Wallet signing key has been loaded.`);
 
 // Then we can send the payment. Note that this isn't paying the invoice we just made because
 // you can't actually pay your own invoice. Let's just pay a pre-existing AMP invoice instead.
 const ampInvoice =
   "lnbcrt1pjr8xwypp5xqj2jfpkz095s8zu57ktsq8vt8yazwcmqpcke9pvl67ne9cpdr0qdqj2a5xzumnwd6hqurswqcqzpgxq9z0rgqsp55hfn0caa5sexea8u979cckkmwelw6h3zpwel5l8tn8s0elgwajss9q8pqqqssqefmmw79tknhl5xhnh7yfepzypxknwr9r4ya7ueqa6vz20axvys8se986hwj6gppeyzst44hm4yl04c4dqjjpqgtt0df254q087sjtfsq35yagj";
-const payment = await client.payInvoiceAndAwaitResult(ampInvoice, 1000, 60, 10);
+const payment = await client.payInvoiceAndAwaitResult(ampInvoice, 1_000_000, 100_000);
 console.log(`Payment done with status= ${payment.status}, ID = ${payment.id}`);
 console.log("");